requirements.txt
fastapi
uvicorn
groq
pydantic
requests
python-dotenv
openai